modification of function names TeX_... -> Tex_... 	cosmetic change, but this is easier to write them menu name: Update -> UpdatePackage 	more clear name (IMO)
Fixed problem with UpdateAll. But here I see one problem. This works
only if in file exists line \begin{document}. This is fundamental.
IMO there is no sense in figuring out all ways of searching preamble.
First 20 lines? If there are rich comments preamble can be much longer.
Whole file? If this is file about latex in body of file can be line with
\usepackage{ and menu will give information about file which noexists in
preamble (and with weird formatting whole function may go to hell).