Implemented Carl's idea of having the log file also being shown when RunLaTeX() is run. This is what happens presently:
When you press \ll, then the quickfix window with one error on each line
opens up. Beneath the quickfix window, the log file is opened in a
"preview" window. The display within the preview window is adjusted to
match with the error currently being examined by the user. When the user
moves around within the quickfix window, this display is automatically
adjusted to keep in sync.

When the user presses <enter> in the quickfix window, then he is taken to
the correct line and column number of the .tex file. The column number is
computed either from the length of the broken line after l.\d\+ in the log
file or if we have a undefined citation, then search for the missing
citation word...

New functions added:

PositionPreviewWindow() : positions the display of the log file
UpdatePreviewWindow() : positions the display of the log file and then
                        returns to the quickfix window.
GotoErrorLocation() : goes to the correct line and column number of the
                      .tex file being edited.

TODO:
1. multiple errors on the same line dont work. the preview window gets
   stuck on the first error.
2. test more robustly.
3. Ask people about various formats of LaTeX warnings, so we can guess
   column numbers in other situations too. As of now, only undefined
   citations and errors are caught well. For all other kinds, we goto the
   first column of the correct line number.