This script provides a command :RemoteOpen which is supposed to be used from the command line as follows:
gvim -c ":RemoteOpen +<lnum> <filename>"

where <lnum> is the line-number you wish <filename> to open to.  What will
happen is that a new gvim will start up and enquire from all previous
sessions if <filename> is already open in any of them. If it is, then it
will edit the file in that session and bring it to the foreground and itself
quit. Otherwise, it will not quit and instead open up the file for editing
at <lnum>.

This was mainly created to be used with Yap (the dvi previewer in miktex),
so you can specify the program for "inverse search" as specified above.
This ensures that the inverse search uses the correct gvim each time.